prototype map update
Build a C# solution for the project.

Installed a plugin called Terrabrush (the C# solution was required for this.)

proto_map is now a thing. this is the new test area

The assets folder has two additional folders called tempfoliage and temptextures. they are as their names imply: placeholder assets for easy map concepts.

I believe those are all the changes made.